Clean up of all dns related functionality. Removal of ACMESharp.
