+++ LetsEncrypt-SiteExtension/Controllers/Api/CertificateController.cs
@@ -83,69 +83,5 @@ namespace LetsEncrypt.SiteExtension.Controllers.Api
 
             return Ok(await mgr.AddCertificate());
         }
-
-        /// <summary>
-        /// Requests a Let's Encrypt certificate using the DNS challenge, using Azure DNS. 
-        /// </summary>
-        /// <param name="model"></param>
-        /// <param name="apiversion"></param>
-        /// <returns></returns>
-        [HttpPost]
-        [Route("api/certificates/challengeprovider/dns/azure")]
-        [ResponseType(typeof(CertificateInstallModel))]
-        public async Task<IHttpActionResult> Generate(DnsAzureModel model, [FromUri(Name = "api-version")]string apiversion = null)
-        {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            var res = await CertificateManager.RequestDnsChallengeCertificate(model.AzureDnsEnvironment, model.AcmeConfig);
-
-            return Ok(res);
-        }
-
-        /// <summary>
-        /// Requests a Let's Encrypt certificate using the DNS challenge, using Azure DNS. 
-        /// </summary>
-        /// <param name="model"></param>
-        /// <param name="apiversion"></param>
-        /// <returns></returns>
-        [HttpPost]
-        [Route("api/certificates/challengeprovider/dns-v2/azure")]
-        [ResponseType(typeof(CertificateInstallModel))]
-        public async Task<IHttpActionResult> Generate_v2(DnsAzureModel model, [FromUri(Name = "api-version")]string apiversion = null)
-        {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            var res = await CertificateManager.RequestDnsChallengeCertificate(model.AzureDnsEnvironment, model.AcmeConfig);
-
-            return Ok(res);
-        }
-
-        /// <summary>
-        /// Requests a Let's Encrypt certificate using the DNS challenge, using Azure DNS. The 
-        /// certificate is installed to the web app. 
-        /// </summary>
-        /// <param name="model"></param>
-        /// <param name="apiversion"></param>
-        /// <returns></returns>
-        [HttpPost]
-        [Route("api/certificates/challengeprovider/dns/azure/certificateinstall/azurewebapp")]
-        [ResponseType(typeof(CertificateInstallModel))]
-        public async Task<IHttpActionResult> GenerateAndInstall(DnsAzureInstallModel model, [FromUri(Name = "api-version")]string apiversion = null)
-        {
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            var mgr = CertificateManager.CreateAzureDnsWebAppCertificateManager(model.AzureWebAppEnvironment, model.AcmeConfig, model.CertificateSettings, model);
-
-            return Ok(await mgr.AddCertificate());
-        }
     }
 }
