+++ LetsEncrypt.SiteExtension.Core/Services/BaseHttpAuthorizationChallengeProvider.cs
@@ -1,17 +1,12 @@
-﻿using ACMESharp;
-using ACMESharp.ACME;
-using Certes;
+﻿using Certes;
 using Certes.Acme;
 using Certes.Acme.Resource;
-using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
-using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 
 namespace LetsEncrypt.Azure.Core.Services
@@ -36,9 +31,7 @@ namespace LetsEncrypt.Azure.Core.Services
     </authorization>
   </system.web>
 </configuration>";
-        private ACMESharp.AcmeClient client;
-        private IOrderContext order;
-
+        
         public virtual Task EnsureDirectory() {
             return Task.CompletedTask;
         }
@@ -52,43 +45,21 @@ namespace LetsEncrypt.Azure.Core.Services
 
         public abstract Task CleanupChallengeFile(string path);
 
-        public void RegisterClient(object client)
-        {
-            if (client is ACMESharp.AcmeClient)
-                this.client = client as ACMESharp.AcmeClient;
-            else if (client as IOrderContext != null)
-                this.order = client as IOrderContext;
-            else
-                throw new InvalidProgramException($"Unable to determine type of AcmeClient {client.GetType()}");
-        }
 
-        public async Task<string> Authorize(List<string> allDnsIdentifiers)
-        {
-            if (client != null)
-            {
-                return await this.AuthorizeOld(allDnsIdentifiers);
-            }
-            if (order != null)
-            {
-                return await this.AuthorizeNew(allDnsIdentifiers);
-            }
-
-            throw new Exception("order and client are both null");
-        }
-
-        private async Task<string> AuthorizeNew(List<string> allDnsIdentifiers)
+        public async Task<string> Authorize(IOrderContext order, List<string> allDnsIdentifiers)
         {
             await EnsureDirectory();
             await EnsureWebConfig();
 
             var auths = await order.Authorizations();
+            var i = 0;
             foreach (var auth in auths)
             {                
 
                 var authz = await auth.Http();
                 var tokenRelativeUri = $".well-known/acme-challenge/{authz.Token}";
                 await PersistsChallengeFile(tokenRelativeUri, authz.KeyAuthz);
-                var challengeUri = new Uri($"http://{allDnsIdentifiers.First()}/{tokenRelativeUri}");
+                var challengeUri = new Uri($"http://{allDnsIdentifiers.ElementAt(i)}/{tokenRelativeUri}");
                 var retry = await ValidateChallengeFile(challengeUri);
                 if (retry == 0)
                     throw new Exception($"Unable to validate presence of http challenge at {challengeUri} ensure that it is browsable");
@@ -109,81 +80,11 @@ namespace LetsEncrypt.Azure.Core.Services
                 {
                     return response.Status.ToString();
                 }
+                i++;
             }
             return "valid";
         }
 
-        public async Task<string> AuthorizeOld(List<string> allDnsIdentifiers)
-        {
-            List<AuthorizationState> authStatus = new List<AuthorizationState>();
-
-            await EnsureDirectory();
-            await EnsureWebConfig();
-
-            foreach (var dnsIdentifier in allDnsIdentifiers)
-            {
-                //var dnsIdentifier = target.Host;                
-                Console.WriteLine($"\nAuthorizing Identifier {dnsIdentifier} Using Challenge Type {AcmeProtocol.CHALLENGE_TYPE_HTTP}");
-                Trace.TraceInformation("Authorizing Identifier {0} Using Challenge Type {1}", dnsIdentifier, AcmeProtocol.CHALLENGE_TYPE_HTTP);
-                var authzState = client.AuthorizeIdentifier(dnsIdentifier);                
-                var challenge = client.DecodeChallenge(authzState, AcmeProtocol.CHALLENGE_TYPE_HTTP);
-                var httpChallenge = challenge.Challenge as HttpChallenge;
-
-                await PersistsChallengeFile(httpChallenge.FilePath, httpChallenge.FileContent);
-                try
-                {
-                    var answerUri = new Uri(httpChallenge.FileUrl);
-                    int retry = await ValidateChallengeFile(answerUri);
-                    Console.WriteLine(" Submitting answer");
-                    Trace.TraceInformation("Submitting answer");
-                    authzState.Challenges = new AuthorizeChallenge[] { challenge };
-                    client.SubmitChallengeAnswer(authzState, AcmeProtocol.CHALLENGE_TYPE_HTTP, true);
-
-                    // have to loop to wait for server to stop being pending. 
-                    retry = 0;
-                    while (authzState.Status == "pending" && retry < 6)
-                    {
-                        retry++;
-                        Console.WriteLine(" Refreshing authorization attempt " + retry);
-                        Trace.TraceInformation("Refreshing authorization attempt " + retry);
-                        await Task.Delay(2000 * retry);  // this has to be here to give ACME server a chance to think
-                        var newAuthzState = client.RefreshIdentifierAuthorization(authzState);
-                        if (newAuthzState.Status != "pending")
-                            authzState = newAuthzState;
-                    }
-
-                    Console.WriteLine($" Authorization Result: {authzState.Status}");
-                    Trace.TraceInformation("Auth Result {0}", authzState.Status);
-                    if (authzState.Status == "invalid" || authzState.Status == "pending")
-                    {
-                        Trace.TraceError("Authorization Failed {0}", authzState.Status);
-                        Trace.TraceInformation("Full Error Details {0}", JsonConvert.SerializeObject(authzState));
-                        Console.WriteLine($"The ACME server was probably unable to reach {answerUri}");
-                        Trace.TraceError("Unable to reach {0}", answerUri);
-                        Console.WriteLine("\nCheck in a browser to see if the answer file is being served correctly.");
-                        throw new Exception($"The Lets Encrypt ACME server was probably unable to reach {answerUri} view error report from Lets Encrypt at {authzState.Uri} for more information");
-                    }
-                    authStatus.Add(authzState);
-                }
-                finally
-                {
-                    if (authzState.Status == "valid")
-                    {
-                        Console.WriteLine(" Deleting answer");
-                        Trace.TraceInformation("Deleting answer");
-                        await CleanupChallengeFile(httpChallenge.FilePath);
-                    }
-                }
-            }
-            foreach (var authState in authStatus)
-            {
-                if (authState.Status != "valid")
-                {
-                    return authState.Status;
-                }
-            }
-            return "valid";
-        }
 
         private static async Task<int> ValidateChallengeFile(Uri answerUri)
         {
