+++ LetsEncrypt.SiteExtension.Core/Services/BlobStorageAuthorizationChallengeProvider.cs
@@ -1,9 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
-using ACMESharp.ACME;
 using Microsoft.WindowsAzure.Storage;
 using Microsoft.WindowsAzure.Storage.Blob;
