+++ LetsEncrypt.SiteExtension.Core/Services/IAuthorizationChallengeProvider.cs
@@ -1,10 +1,4 @@
-﻿using ACMESharp;
-using Certes;
-using LetsEncrypt.Azure.Core.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
+﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 
 namespace LetsEncrypt.Azure.Core.Services
@@ -16,7 +10,6 @@ namespace LetsEncrypt.Azure.Core.Services
         /// </summary>
         /// <param name="dnsIdentifiers"></param>
         /// <returns></returns>
-        Task<string> Authorize(List<string> dnsIdentifiers);
-        void RegisterClient(object client);
+        Task<string> Authorize(Certes.Acme.IOrderContext context, List<string> dnsIdentifiers);    
     }
 }
